+++ src/constants/lifecycle.ts
@@ -19,6 +19,7 @@ export const GwCoreLifecycleAction = Object.freeze({
   Transaction: '_transaction',
   TransactionResponse: '_transaction-response',
   Disable: '_disable',
+  Enable: '_enable',
   Purge: '_purge',
   Status: '_status',
   DisableDescendants: '_disable-descendants',
@@ -45,6 +46,7 @@ export const GwCoreLifecycleRequestType = Object.freeze({
   IndividualOrganizationPurge: 'Family-purge-request-v1.0',
   IndividualMemberPurge: LifecycleRequestType.RelatedPersonPurge,
   TenantStatus: 'Organization-lifecycle-status-request-v1.0',
+  TenantEnable: 'Organization-enable-request-v1.0',
   TenantDisableDescendants: 'Organization-disable-descendants-request-v1.0',
   TenantPurgeDescendants: 'Organization-purge-descendants-request-v1.0',
 } as const);